multiprocessing + diskcheck
